Return a sensible result irrespective of input size
At least 2 values are necessary to compute a result, though, far
more are needed to return a useful result. We must deal with the
possibility of n = 0 or 1.
There are two options here:
1) As implemented in this commit: return a worthless result, but do
not kill the program.
2) Kill the program, i.e. replace the branch with `assert!(n > 1)`

My preference is 2), as I am of the opinion that invalid inputs
to a numerical program should result in said program dying as
early as possible, as nothing useful can be derived from garbage.
On the other hand, it may be friendlier to avoid terminating users'
programs.